+++ src/app/locales/en/components/-storage/s3.js
@@ -7,6 +7,14 @@ export default {
   secretKey: { name: 'Admin secret key' },
   hostname: { name: 'Hostname' },
   bucketName: { name: 'Bucket name' },
+  verifyServerCertificate: {
+    name: 'Verify server cert.',
+    tip: 'Enables or disables verification of the S3 server SSL certificate.',
+  },
+  region: {
+    name: 'Region',
+    tip: 'Allows to specify a custom S3 region, which will be send with each request to the S3 server.',
+  },
   blockSize: {
     name: 'Block size [bytes]',
     tip: 'Storage block size in bytes. In case the block size is "0" and "canonical" path type is selected, each file is stored in a single S3 object. This value must be set to "0" to enable data import from an existing S3 bucket.',