+++ src/app/locales/en/services/storage-actions.js
@@ -1,7 +1,4 @@
 export default {
-  storageModifiedSuccessfully: 'Storage backend "{{name}}" has been modified successfully.',
-  modifyingStorage: 'modifying storage backend',
-  storageCheckFailed: 'File read/write test has failed after "{{name}}" storage backend modification. Please make sure that the storage backend configuration is correct.',
   storageRemovedSuccessfully: 'Storage backend "{{name}}" has been removed successfully.',
   removingStorage: 'removing storage backend',
 };