+++ src/app/locales/en/utils/storage-actions.js
@@ -1,0 +1,5 @@
+import saveStorageModificationAction from './storage-actions/save-storage-modification-action';
+
+export default {
+  saveStorageModificationAction,
+};